Stat form

Closes #4, #5, #7, and #8

See merge request medphys_funcimag/web!6

+++ forms.py
@@ -1,5 +1,5 @@
 from flask_wtf import FlaskForm
-from wtforms import StringField, PasswordField, BooleanField, SubmitField, SelectField, FloatField
+from wtforms import StringField, PasswordField, BooleanField, SubmitField, SelectField, FloatField, IntegerField
 from wtforms.validators import DataRequired
 
 class LoginForm(FlaskForm):
@@ -11,7 +11,7 @@ class LoginForm(FlaskForm):
 
 class powerForm(FlaskForm):
     effect = FloatField('Effect', default=None , validators=[DataRequired()])
-    nobs = FloatField('nobs', default=None, validators=[DataRequired()]) # nobs = number of observations
+    nobs = IntegerField('nobs', default=None, validators=[DataRequired()]) # nobs = number of observations
     alpha = SelectField('Alpha', choices=[(0.05, '0.05'), (0.01, '0.01')]) 
     power = FloatField('Power', default=0.8, validators=[DataRequired()])
     test = SelectField('Test', choices=[('independent', 'independent samples'), ('paired', 'paired samples')])
